Update tests, add more socket and ping tests, add a coverage badge
