+++ packages/client-node/src/index.ts
@@ -19,6 +19,11 @@ export {
   type ExecResult,
   type InsertResult,
   type DataFormat,
+  type RawDataFormat,
+  type JSONDataFormat,
+  type StreamableDataFormat,
+  type StreamableJSONDataFormat,
+  type SingleDocumentJSONFormat,
   type Logger,
   type LogParams,
   type ErrorLogParams,
